feat(output): add all-toggle + done btn per domain
This commit adds a 'Done' link to hide all emails from a particular
domain in the links column of the HTML output, and also adds a link back
to the official GitHub page in the output page header.

The binaries now also include the version tag.

The documentation has been updated to demonstrate how users of email
providers that don't allow eml archive exports can use Thunderbird or
similar email clients to save their own emails as eml files for use with
unsubscan.